Fixed menues add global var menues for View and add categories

+++ database/migrations/2025_06_21_094910_create_banners_table.php
@@ -33,6 +33,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('banners');
+        if (Schema::hasTable("banners")) {
+            Schema::dropIfExists('banners');
+        }
     }
 };
